+++ robotech-api/src/main/java/rebue/robotech/dic/DicUtils.java
@@ -12,12 +12,12 @@ public class DicUtils {
      * 外面的Map是缓存所有字典类的集合，key为类名，value为字典类所有值的集合
      * 里面的Map是缓存该字典类所有值的集合，key为字典的code，value为字典的值
      */
-    private static Map<String, Map<Byte, Dic>> caches = new ConcurrentHashMap<>();
+    private static Map<String, Map<Integer, Dic>> caches = new ConcurrentHashMap<>();
 
     /**
      * 获取字典类的所有值的集合
      */
-    public static Map<Byte, Dic> getItems(final Class<?> dicClass) {
+    public static Map<Integer, Dic> getItems(final Class<?> dicClass) {
         if (!dicClass.isEnum()) {
             throw new IllegalArgumentException("参数必须是枚举类型");
         }
@@ -27,7 +27,7 @@ public class DicUtils {
         }
 
         // 先从缓存中查找
-        Map<Byte, Dic> items = caches.get(dicClass.getName());
+        Map<Integer, Dic> items = caches.get(dicClass.getName());
         if (items != null) {
             return items;
         }
@@ -47,7 +47,7 @@ public class DicUtils {
      * @param dicClass 字典类型
      * @param code     字典的编码
      */
-    public static Dic getItem(final Class<?> dicClass, final byte code) {
+    public static Dic getItem(final Class<?> dicClass, final Integer code) {
         return getItems(dicClass).get(code);
     }
 
@@ -57,7 +57,7 @@ public class DicUtils {
      * @param dicClass 字典类型
      * @param code     字典的编码
      */
-    public static boolean isValid(final Class<?> dicClass, final byte code) {
+    public static boolean isValid(final Class<?> dicClass, final Integer code) {
         return getItem(dicClass, code) != null;
     }
 }
